// src/services/Reline/interfaces/node.relineSchema.ts
import { RelineElementBaseSchema } from "./elementBase.relineSchema";
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface RelineNodeSchema extends RelineElementBaseSchema{
  id: string;
  metadata: {
    basic: {
      id: string;
      [k: string]: unknown;
    };
    [k: string]: {
      [k: string]: unknown;
    };
  };
  [k: string]: unknown;
}